move package in own directory, split it up in files
